Create Intersection_in_linkedlist.py
In this code, I will determine whether there is an intersection point in the linked list or not.

The above code has a complexity of O (n).

The steps to follow are:

Step1: Find the length of both the linked lists.

Step2: Find the difference in the length of the linked lists.

Step3:Move the pointer forward by the difference.

Step4: In the other loop, find whether they meet at a particular point or not. If yes (intersection point exists), else not.